refactor: move around existing manifest-related code
