+++ internal/backup_restore/manifest.go
@@ -3,6 +3,8 @@ package backup_restore
 import (
 	"fmt"
 	"time"
+
+	"github.com/influxdata/influx-cli/v2/api"
 )
 
 type FileCompression int
@@ -35,9 +37,13 @@ func (c FileCompression) String() string {
 	}
 }
 
-const ManifestExtension = "manifest"
+const (
+	ManifestExtension = "manifest"
+	ManifestVersion   = 2
+)
 
 type Manifest struct {
+	Version int                   `json:"manifestVersion"`
 	KV      ManifestFileEntry     `json:"kv"`
 	SQL     *ManifestFileEntry    `json:"sql,omitempty"`
 	Buckets []ManifestBucketEntry `json:"buckets"`
@@ -91,4 +97,104 @@ type ManifestSubscription struct {
 	Name         string   `json:"name"`
 	Mode         string   `json:"mode"`
 	Destinations []string `json:"destinations"`
+}
+
+func ConvertBucketManifest(manifest api.BucketMetadataManifest, getShard func(shardId int64) (*ManifestFileEntry, error)) (ManifestBucketEntry, error) {
+	m := ManifestBucketEntry{
+		OrganizationID:         manifest.OrganizationID,
+		OrganizationName:       manifest.OrganizationName,
+		BucketID:               manifest.BucketID,
+		BucketName:             manifest.BucketName,
+		Description:            manifest.Description,
+		DefaultRetentionPolicy: manifest.DefaultRetentionPolicy,
+		RetentionPolicies:      make([]ManifestRetentionPolicy, len(manifest.RetentionPolicies)),
+	}
+
+	for i, rp := range manifest.RetentionPolicies {
+		var err error
+		m.RetentionPolicies[i], err = ConvertRetentionPolicy(rp, getShard)
+		if err != nil {
+			return ManifestBucketEntry{}, err
+		}
+	}
+
+	return m, nil
+}
+
+func ConvertRetentionPolicy(manifest api.RetentionPolicyManifest, getShard func(shardId int64) (*ManifestFileEntry, error)) (ManifestRetentionPolicy, error) {
+	m := ManifestRetentionPolicy{
+		Name:               manifest.Name,
+		ReplicaN:           manifest.ReplicaN,
+		Duration:           manifest.Duration,
+		ShardGroupDuration: manifest.ShardGroupDuration,
+		ShardGroups:        make([]ManifestShardGroup, len(manifest.ShardGroups)),
+		Subscriptions:      make([]ManifestSubscription, len(manifest.Subscriptions)),
+	}
+
+	for i, sg := range manifest.ShardGroups {
+		var err error
+		m.ShardGroups[i], err = ConvertShardGroup(sg, getShard)
+		if err != nil {
+			return ManifestRetentionPolicy{}, err
+		}
+	}
+
+	for i, s := range manifest.Subscriptions {
+		m.Subscriptions[i] = ManifestSubscription{
+			Name:         s.Name,
+			Mode:         s.Mode,
+			Destinations: s.Destinations,
+		}
+	}
+
+	return m, nil
+}
+
+func ConvertShardGroup(manifest api.ShardGroupManifest, getShard func(shardId int64) (*ManifestFileEntry, error)) (ManifestShardGroup, error) {
+	m := ManifestShardGroup{
+		ID:          manifest.Id,
+		StartTime:   manifest.StartTime,
+		EndTime:     manifest.EndTime,
+		DeletedAt:   manifest.DeletedAt,
+		TruncatedAt: manifest.TruncatedAt,
+		Shards:      make([]ManifestShardEntry, 0, len(manifest.Shards)),
+	}
+
+	for _, sh := range manifest.Shards {
+		maybeShard, err := ConvertShard(sh, getShard)
+		if err != nil {
+			return ManifestShardGroup{}, err
+		}
+		// Shard deleted mid-backup.
+		if maybeShard == nil {
+			continue
+		}
+		m.Shards = append(m.Shards, *maybeShard)
+	}
+
+	return m, nil
+}
+
+func ConvertShard(manifest api.ShardManifest, getShard func(shardId int64) (*ManifestFileEntry, error)) (*ManifestShardEntry, error) {
+	shardFileInfo, err := getShard(manifest.Id)
+	if err != nil {
+		return nil, fmt.Errorf("failed to download snapshot of shard %d: %w", manifest.Id, err)
+	}
+	if shardFileInfo == nil {
+		return nil, nil
+	}
+
+	m := ManifestShardEntry{
+		ID:                manifest.Id,
+		ShardOwners:       make([]ShardOwnerEntry, len(manifest.ShardOwners)),
+		ManifestFileEntry: *shardFileInfo,
+	}
+
+	for i, o := range manifest.ShardOwners {
+		m.ShardOwners[i] = ShardOwnerEntry{
+			NodeID: o.NodeID,
+		}
+	}
+
+	return &m, nil
 }
